+++ core/app/Http/Controllers/UserController.php
@@ -1,6 +1,8 @@
 <?php
 
 namespace App\Http\Controllers;
+
+use App\models\Penjualan;
 use App\User;
 use Illuminate\Http\Request;
 
@@ -46,8 +48,14 @@ class UserController extends Controller
     public function destroy(Request $request)
     {
         $user = User::find($request->id);
-        $user->delete();
-        return response()->json($user);
+        $cek = Penjualan::where('user_id',$user->id)->count();
+        if($cek == 0){
+            $user->delete();
+            return response()->json('oke');
+        }else{
+            return response()->json('no');
+        }
+        
     }
 
     public function update(Request $request, $id)
